Improvements to event system. Replace string keys with classes.
